requirements.txt
asgiref==3.7.2
channels==4.0.0
Django==4.2.5
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
uuid==1.30
